Moving things around; new test_slapchop.bat script (which doesn't do much yet.)
