+++ src/test/__init__.py
@@ -1,0 +1,3 @@
+from .test_basic_input_functions import *
+from .test_message_responder_statics import *
+from .test_spoofer_resources import *
