+++ src/test/framework/spoof_factory.py
