+++ test_slapchop.bat
@@ -1,0 +1,1 @@
+python3 src/test_all.py
